old-tests/submission/W3C/domhtml/HTMLMenuElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - MENU</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-menu-compact" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<menu compact="COMPACT">
<li>Interview</li>
<li>Paperwork</li>
<li>Give start date</li>
</menu>



<script>
var domtest = async_test("The compact IDL attribute of the menu element must reflect the content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcompact;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("menu");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vcompact = testNode.compact;

      assert_true(vcompact);

    domtest.done();
}));

</script>
<p>The compact IDL attribute of the menu element must reflect the content attribute of the same name.</p>
<div id="log"></div>
</body></html>
